+++ packages/building-rollup/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.22](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.9.21...@open-wc/building-rollup@0.9.22) (2019-08-25)
+
+**Note:** Version bump only for package @open-wc/building-rollup
+
+
+
+
+
+## [0.9.21](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.9.20...@open-wc/building-rollup@0.9.21) (2019-08-25)
+
+**Note:** Version bump only for package @open-wc/building-rollup
+
+
+
+
+
 ## [0.9.20](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.9.19...@open-wc/building-rollup@0.9.20) (2019-08-22)
 
 **Note:** Version bump only for package @open-wc/building-rollup